+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2026.05.1
+
+### Added
+
+- Driver for parsing command line arguments
+- .clangd file
+
+### Changed
+
+- Case style of functions in .clang-tidy (now it is CamelCase)
+- Removed 'k' prefix for constants in .clang-tidy
+- Disable `cert-err58-cpp` rule in .clang-tidy
+- Added include_directories for include/ in the root of project
+
 ## [0.0.1] - 2026.05.01
 
 ### Added